+++ android-mic-rs/src/main.rs
@@ -9,9 +9,7 @@ extern crate log;
 mod app;
 mod audio;
 mod config;
-mod message;
 mod streamer;
-// mod user_action;
 mod my_widget;
 mod utils;
 mod view;
