--- android-mic-rs/src/message.rs
@@ -1,1 +1,0 @@
-
