+++ android-mic-rs/src/my_widget/mod.rs
@@ -1,6 +1,1 @@
-// pub mod drop_down;
-// pub use drop_down::drop_down;
 pub mod audio_wave;
-
-// pub mod alignment;
-// pub mod offset;
